@@ -49,49 +49,49 @@ const html = `<p>English auction for NFT.</p>
 
     <span class="hljs-keyword">address</span> <span class="hljs-keyword">public</span> highestBidder;
     <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> highestBid;
-    <span class="hljs-keyword">mapping</span>(<span class="hljs-keyword">address</span> =&gt; <span class="hljs-keyword">uint</span>) <span class="hljs-keyword">public</span> bids;
+    <span class="hljs-keyword">mapping</span>(<span class="hljs-keyword">address</span> <span class="hljs-operator">=</span><span class="hljs-operator">&gt;</span> <span class="hljs-keyword">uint</span>) <span class="hljs-keyword">public</span> bids;
 
     <span class="hljs-function"><span class="hljs-keyword">constructor</span>(<span class="hljs-params">
         <span class="hljs-keyword">address</span> _nft,
         <span class="hljs-keyword">uint</span> _nftId,
         <span class="hljs-keyword">uint</span> _startingBid
     </span>) </span>{
-        nft = IERC721(_nft);
-        nftId = _nftId;
+        nft <span class="hljs-operator">=</span> IERC721(_nft);
+        nftId <span class="hljs-operator">=</span> _nftId;
 
-        seller = <span class="hljs-keyword">payable</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>);
-        highestBid = _startingBid;
+        seller <span class="hljs-operator">=</span> <span class="hljs-keyword">payable</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>);
+        highestBid <span class="hljs-operator">=</span> _startingBid;
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">start</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">external</span></span> </span>{
-        <span class="hljs-built_in">require</span>(!started, <span class="hljs-string">"started"</span>);
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> == seller, <span class="hljs-string">"not seller"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-operator">!</span>started, <span class="hljs-string">"started"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> seller, <span class="hljs-string">"not seller"</span>);
 
         nft.transferFrom(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>, <span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>), nftId);
-        started = <span class="hljs-literal">true</span>;
-        endAt = <span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> + <span class="hljs-number">7</span> <span class="hljs-literal">days</span>;
+        started <span class="hljs-operator">=</span> <span class="hljs-literal">true</span>;
+        endAt <span class="hljs-operator">=</span> <span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> <span class="hljs-operator">+</span> <span class="hljs-number">7</span> <span class="hljs-literal">days</span>;
 
         <span class="hljs-keyword">emit</span> Start();
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">bid</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">external</span></span> <span class="hljs-title"><span class="hljs-keyword">payable</span></span> </span>{
         <span class="hljs-built_in">require</span>(started, <span class="hljs-string">"not started"</span>);
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> &lt; endAt, <span class="hljs-string">"ended"</span>);
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> &gt; highestBid, <span class="hljs-string">"value &lt; highest"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> <span class="hljs-operator">&lt;</span> endAt, <span class="hljs-string">"ended"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> <span class="hljs-operator">&gt;</span> highestBid, <span class="hljs-string">"value &lt; highest"</span>);
 
-        <span class="hljs-keyword">if</span> (highestBidder != <span class="hljs-keyword">address</span>(<span class="hljs-number">0</span>)) {
-            bids[highestBidder] += highestBid;
+        <span class="hljs-keyword">if</span> (highestBidder <span class="hljs-operator">!</span><span class="hljs-operator">=</span> <span class="hljs-keyword">address</span>(<span class="hljs-number">0</span>)) {
+            bids[highestBidder] <span class="hljs-operator">+</span><span class="hljs-operator">=</span> highestBid;
         }
 
-        highestBidder = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
-        highestBid = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span>;
+        highestBidder <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
+        highestBid <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span>;
 
         <span class="hljs-keyword">emit</span> Bid(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>, <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span>);
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">withdraw</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">external</span></span> </span>{
-        <span class="hljs-keyword">uint</span> bal = bids[<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>];
-        bids[<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>] = <span class="hljs-number">0</span>;
+        <span class="hljs-keyword">uint</span> bal <span class="hljs-operator">=</span> bids[<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>];
+        bids[<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>] <span class="hljs-operator">=</span> <span class="hljs-number">0</span>;
         <span class="hljs-keyword">payable</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>).<span class="hljs-built_in">transfer</span>(bal);
 
         <span class="hljs-keyword">emit</span> Withdraw(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>, bal);
@@ -99,11 +99,11 @@ const html = `<p>English auction for NFT.</p>
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">end</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">external</span></span> </span>{
         <span class="hljs-built_in">require</span>(started, <span class="hljs-string">"not started"</span>);
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> &gt;= endAt, <span class="hljs-string">"not ended"</span>);
-        <span class="hljs-built_in">require</span>(!ended, <span class="hljs-string">"ended"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span> <span class="hljs-operator">&gt;</span><span class="hljs-operator">=</span> endAt, <span class="hljs-string">"not ended"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-operator">!</span>ended, <span class="hljs-string">"ended"</span>);
 
-        ended = <span class="hljs-literal">true</span>;
-        <span class="hljs-keyword">if</span> (highestBidder != <span class="hljs-keyword">address</span>(<span class="hljs-number">0</span>)) {
+        ended <span class="hljs-operator">=</span> <span class="hljs-literal">true</span>;
+        <span class="hljs-keyword">if</span> (highestBidder <span class="hljs-operator">!</span><span class="hljs-operator">=</span> <span class="hljs-keyword">address</span>(<span class="hljs-number">0</span>)) {
             nft.safeTransferFrom(<span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>), highestBidder, nftId);
             seller.<span class="hljs-built_in">transfer</span>(highestBid);
         } <span class="hljs-keyword">else</span> {
