@@ -9,13 +9,13 @@ const html = `<p>Transactions are paid with <code>ether</code>.</p>
 <span class="hljs-meta"><span class="hljs-keyword">pragma</span> <span class="hljs-keyword">solidity</span> ^0.8.10;</span>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">EtherUnits</span> </span>{
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> oneWei = <span class="hljs-number">1</span> <span class="hljs-literal">wei</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> oneWei <span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">wei</span>;
     <span class="hljs-comment">// 1 wei is equal to 1</span>
-    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> isOneWei = <span class="hljs-number">1</span> <span class="hljs-literal">wei</span> == <span class="hljs-number">1</span>;
+    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> isOneWei <span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">wei</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">1</span>;
 
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> oneEther = <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> oneEther <span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>;
     <span class="hljs-comment">// 1 ether is equal to 10^18 wei</span>
-    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> isOneEther = <span class="hljs-number">1</span> <span class="hljs-literal">ether</span> == <span class="hljs-number">1e18</span>;
+    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> isOneEther <span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">ether</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">1e18</span>;
 }
 </code></pre>
 `
