@@ -29,24 +29,24 @@ Now no one can deposit and the winner cannot be set.
 */</span>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">EtherGame</span> </span>{
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> targetAmount = <span class="hljs-number">7</span> <span class="hljs-literal">ether</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> targetAmount <span class="hljs-operator">=</span> <span class="hljs-number">7</span> <span class="hljs-literal">ether</span>;
     <span class="hljs-keyword">address</span> <span class="hljs-keyword">public</span> winner;
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">deposit</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> <span class="hljs-title"><span class="hljs-keyword">payable</span></span> </span>{
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> == <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>, <span class="hljs-string">"You can only send 1 Ether"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>, <span class="hljs-string">"You can only send 1 Ether"</span>);
 
-        <span class="hljs-keyword">uint</span> balance = <span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>).<span class="hljs-built_in">balance</span>;
-        <span class="hljs-built_in">require</span>(balance &lt;= targetAmount, <span class="hljs-string">"Game is over"</span>);
+        <span class="hljs-keyword">uint</span> balance <span class="hljs-operator">=</span> <span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>).<span class="hljs-built_in">balance</span>;
+        <span class="hljs-built_in">require</span>(balance <span class="hljs-operator">&lt;</span><span class="hljs-operator">=</span> targetAmount, <span class="hljs-string">"Game is over"</span>);
 
-        <span class="hljs-keyword">if</span> (balance == targetAmount) {
-            winner = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
+        <span class="hljs-keyword">if</span> (balance <span class="hljs-operator">=</span><span class="hljs-operator">=</span> targetAmount) {
+            winner <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
         }
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">claimReward</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> </span>{
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> == winner, <span class="hljs-string">"Not winner"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> winner, <span class="hljs-string">"Not winner"</span>);
 
-        (<span class="hljs-keyword">bool</span> sent, ) = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>.<span class="hljs-built_in">call</span>{<span class="hljs-built_in">value</span>: <span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>).<span class="hljs-built_in">balance</span>}(<span class="hljs-string">""</span>);
+        (<span class="hljs-keyword">bool</span> sent, ) <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>.<span class="hljs-built_in">call</span>{<span class="hljs-built_in">value</span>: <span class="hljs-keyword">address</span>(<span class="hljs-built_in">this</span>).<span class="hljs-built_in">balance</span>}(<span class="hljs-string">""</span>);
         <span class="hljs-built_in">require</span>(sent, <span class="hljs-string">"Failed to send Ether"</span>);
     }
 }
@@ -55,7 +55,7 @@ Now no one can deposit and the winner cannot be set.
     EtherGame etherGame;
 
     <span class="hljs-function"><span class="hljs-keyword">constructor</span>(<span class="hljs-params">EtherGame _etherGame</span>) </span>{
-        etherGame = EtherGame(_etherGame);
+        etherGame <span class="hljs-operator">=</span> EtherGame(_etherGame);
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">attack</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> <span class="hljs-title"><span class="hljs-keyword">payable</span></span> </span>{
@@ -63,7 +63,7 @@ Now no one can deposit and the winner cannot be set.
         <span class="hljs-comment">// the game balance &gt;= 7 ether</span>
 
         <span class="hljs-comment">// cast address to payable</span>
-        <span class="hljs-keyword">address</span> <span class="hljs-keyword">payable</span> addr = <span class="hljs-keyword">payable</span>(<span class="hljs-keyword">address</span>(etherGame));
+        <span class="hljs-keyword">address</span> <span class="hljs-keyword">payable</span> addr <span class="hljs-operator">=</span> <span class="hljs-keyword">payable</span>(<span class="hljs-keyword">address</span>(etherGame));
         <span class="hljs-built_in">selfdestruct</span>(addr);
     }
 }
@@ -74,25 +74,25 @@ Now no one can deposit and the winner cannot be set.
 <span class="hljs-meta"><span class="hljs-keyword">pragma</span> <span class="hljs-keyword">solidity</span> ^0.8.10;</span>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">EtherGame</span> </span>{
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> targetAmount = <span class="hljs-number">3</span> <span class="hljs-literal">ether</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> targetAmount <span class="hljs-operator">=</span> <span class="hljs-number">3</span> <span class="hljs-literal">ether</span>;
     <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> balance;
     <span class="hljs-keyword">address</span> <span class="hljs-keyword">public</span> winner;
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">deposit</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> <span class="hljs-title"><span class="hljs-keyword">payable</span></span> </span>{
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> == <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>, <span class="hljs-string">"You can only send 1 Ether"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">1</span> <span class="hljs-literal">ether</span>, <span class="hljs-string">"You can only send 1 Ether"</span>);
 
-        balance += <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span>;
-        <span class="hljs-built_in">require</span>(balance &lt;= targetAmount, <span class="hljs-string">"Game is over"</span>);
+        balance <span class="hljs-operator">+</span><span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">value</span>;
+        <span class="hljs-built_in">require</span>(balance <span class="hljs-operator">&lt;</span><span class="hljs-operator">=</span> targetAmount, <span class="hljs-string">"Game is over"</span>);
 
-        <span class="hljs-keyword">if</span> (balance == targetAmount) {
-            winner = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
+        <span class="hljs-keyword">if</span> (balance <span class="hljs-operator">=</span><span class="hljs-operator">=</span> targetAmount) {
+            winner <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>;
         }
     }
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">claimReward</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> </span>{
-        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> == winner, <span class="hljs-string">"Not winner"</span>);
+        <span class="hljs-built_in">require</span>(<span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span> <span class="hljs-operator">=</span><span class="hljs-operator">=</span> winner, <span class="hljs-string">"Not winner"</span>);
 
-        (<span class="hljs-keyword">bool</span> sent, ) = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>.<span class="hljs-built_in">call</span>{<span class="hljs-built_in">value</span>: balance}(<span class="hljs-string">""</span>);
+        (<span class="hljs-keyword">bool</span> sent, ) <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>.<span class="hljs-built_in">call</span>{<span class="hljs-built_in">value</span>: balance}(<span class="hljs-string">""</span>);
         <span class="hljs-built_in">require</span>(sent, <span class="hljs-string">"Failed to send Ether"</span>);
     }
 }
