@@ -37,12 +37,12 @@ const html = `<p><code>keccak256</code> computes the Keccak-256 hash of the inpu
 }
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">GuessTheMagicWord</span> </span>{
-    <span class="hljs-keyword">bytes32</span> <span class="hljs-keyword">public</span> answer =
+    <span class="hljs-keyword">bytes32</span> <span class="hljs-keyword">public</span> answer <span class="hljs-operator">=</span>
         <span class="hljs-number">0x60298f78cc0b47170ba79c10aa3851d7648bd96f2f8e46a19dbc777c36fb0c00</span>;
 
     <span class="hljs-comment">// Magic word is "Solidity"</span>
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">guess</span>(<span class="hljs-params"><span class="hljs-keyword">string</span> <span class="hljs-keyword">memory</span> _word</span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> <span class="hljs-title"><span class="hljs-keyword">view</span></span> <span class="hljs-title"><span class="hljs-keyword">returns</span></span> (<span class="hljs-params"><span class="hljs-keyword">bool</span></span>) </span>{
-        <span class="hljs-keyword">return</span> <span class="hljs-built_in">keccak256</span>(<span class="hljs-built_in">abi</span>.<span class="hljs-built_in">encodePacked</span>(_word)) == answer;
+        <span class="hljs-keyword">return</span> <span class="hljs-built_in">keccak256</span>(<span class="hljs-built_in">abi</span>.<span class="hljs-built_in">encodePacked</span>(_word)) <span class="hljs-operator">=</span><span class="hljs-operator">=</span> answer;
     }
 }
 </code></pre>
