@@ -9,9 +9,9 @@ const html = `<p>Solidity support conditional statements <code>if</code>, <code>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">IfElse</span> </span>{
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">foo</span>(<span class="hljs-params"><span class="hljs-keyword">uint</span> x</span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> <span class="hljs-title"><span class="hljs-keyword">pure</span></span> <span class="hljs-title"><span class="hljs-keyword">returns</span></span> (<span class="hljs-params"><span class="hljs-keyword">uint</span></span>) </span>{
-        <span class="hljs-keyword">if</span> (x &lt; <span class="hljs-number">10</span>) {
+        <span class="hljs-keyword">if</span> (x <span class="hljs-operator">&lt;</span> <span class="hljs-number">10</span>) {
             <span class="hljs-keyword">return</span> <span class="hljs-number">0</span>;
-        } <span class="hljs-keyword">else</span> <span class="hljs-keyword">if</span> (x &lt; <span class="hljs-number">20</span>) {
+        } <span class="hljs-keyword">else</span> <span class="hljs-keyword">if</span> (x <span class="hljs-operator">&lt;</span> <span class="hljs-number">20</span>) {
             <span class="hljs-keyword">return</span> <span class="hljs-number">1</span>;
         } <span class="hljs-keyword">else</span> {
             <span class="hljs-keyword">return</span> <span class="hljs-number">2</span>;
@@ -25,7 +25,7 @@ const html = `<p>Solidity support conditional statements <code>if</code>, <code>
         <span class="hljs-comment">// return 2;</span>
 
         <span class="hljs-comment">// shorthand way to write if / else statement</span>
-        <span class="hljs-keyword">return</span> _x &lt; <span class="hljs-number">10</span> ? <span class="hljs-number">1</span> : <span class="hljs-number">2</span>;
+        <span class="hljs-keyword">return</span> _x <span class="hljs-operator">&lt;</span> <span class="hljs-number">10</span> ? <span class="hljs-number">1</span> : <span class="hljs-number">2</span>;
     }
 }
 </code></pre>
