@@ -12,12 +12,12 @@ const html = `<p>Solidity supports <code>for</code>, <code>while</code>, and <co
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">Loop</span> </span>{
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">loop</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> </span>{
         <span class="hljs-comment">// for loop</span>
-        <span class="hljs-keyword">for</span> (<span class="hljs-keyword">uint</span> i = <span class="hljs-number">0</span>; i &lt; <span class="hljs-number">10</span>; i++) {
-            <span class="hljs-keyword">if</span> (i == <span class="hljs-number">3</span>) {
+        <span class="hljs-keyword">for</span> (<span class="hljs-keyword">uint</span> i <span class="hljs-operator">=</span> <span class="hljs-number">0</span>; i <span class="hljs-operator">&lt;</span> <span class="hljs-number">10</span>; i<span class="hljs-operator">+</span><span class="hljs-operator">+</span>) {
+            <span class="hljs-keyword">if</span> (i <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">3</span>) {
                 <span class="hljs-comment">// Skip to next iteration with continue</span>
                 <span class="hljs-keyword">continue</span>;
             }
-            <span class="hljs-keyword">if</span> (i == <span class="hljs-number">5</span>) {
+            <span class="hljs-keyword">if</span> (i <span class="hljs-operator">=</span><span class="hljs-operator">=</span> <span class="hljs-number">5</span>) {
                 <span class="hljs-comment">// Exit loop with break</span>
                 <span class="hljs-keyword">break</span>;
             }
@@ -25,8 +25,8 @@ const html = `<p>Solidity supports <code>for</code>, <code>while</code>, and <co
 
         <span class="hljs-comment">// while loop</span>
         <span class="hljs-keyword">uint</span> j;
-        <span class="hljs-keyword">while</span> (j &lt; <span class="hljs-number">10</span>) {
-            j++;
+        <span class="hljs-keyword">while</span> (j <span class="hljs-operator">&lt;</span> <span class="hljs-number">10</span>) {
+            j<span class="hljs-operator">+</span><span class="hljs-operator">+</span>;
         }
     }
 }
