@@ -14,7 +14,7 @@ const html = `<p>Here we introduce you to some primitive data types available in
 <span class="hljs-meta"><span class="hljs-keyword">pragma</span> <span class="hljs-keyword">solidity</span> ^0.8.10;</span>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">Primitives</span> </span>{
-    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> boo = <span class="hljs-literal">true</span>;
+    <span class="hljs-keyword">bool</span> <span class="hljs-keyword">public</span> boo <span class="hljs-operator">=</span> <span class="hljs-literal">true</span>;
 
     <span class="hljs-comment">/*
     uint stands for unsigned integer, meaning non negative integers
@@ -24,9 +24,9 @@ const html = `<p>Here we introduce you to some primitive data types available in
         ...
         uint256 ranges from 0 to 2 ** 256 - 1
     */</span>
-    <span class="hljs-keyword">uint8</span> <span class="hljs-keyword">public</span> u8 = <span class="hljs-number">1</span>;
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u256 = <span class="hljs-number">456</span>;
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u = <span class="hljs-number">123</span>; <span class="hljs-comment">// uint is an alias for uint256</span>
+    <span class="hljs-keyword">uint8</span> <span class="hljs-keyword">public</span> u8 <span class="hljs-operator">=</span> <span class="hljs-number">1</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u256 <span class="hljs-operator">=</span> <span class="hljs-number">456</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u <span class="hljs-operator">=</span> <span class="hljs-number">123</span>; <span class="hljs-comment">// uint is an alias for uint256</span>
 
     <span class="hljs-comment">/*
     Negative numbers are allowed for int types.
@@ -35,15 +35,15 @@ const html = `<p>Here we introduce you to some primitive data types available in
     int256 ranges from -2 ** 255 to 2 ** 255 - 1
     int128 ranges from -2 ** 127 to 2 ** 127 - 1
     */</span>
-    <span class="hljs-keyword">int8</span> <span class="hljs-keyword">public</span> i8 = <span class="hljs-number">-1</span>;
-    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i256 = <span class="hljs-number">456</span>;
-    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i = <span class="hljs-number">-123</span>; <span class="hljs-comment">// int is same as int256</span>
+    <span class="hljs-keyword">int8</span> <span class="hljs-keyword">public</span> i8 <span class="hljs-operator">=</span> <span class="hljs-number">-1</span>;
+    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i256 <span class="hljs-operator">=</span> <span class="hljs-number">456</span>;
+    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i <span class="hljs-operator">=</span> <span class="hljs-number">-123</span>; <span class="hljs-comment">// int is same as int256</span>
 
     <span class="hljs-comment">// minimum and maximum of int</span>
-    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> minInt = <span class="hljs-built_in">type</span>(<span class="hljs-keyword">int</span>).<span class="hljs-built_in">min</span>;
-    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> maxInt = <span class="hljs-built_in">type</span>(<span class="hljs-keyword">int</span>).<span class="hljs-built_in">max</span>;
+    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> minInt <span class="hljs-operator">=</span> <span class="hljs-keyword">type</span>(<span class="hljs-keyword">int</span>).<span class="hljs-built_in">min</span>;
+    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> maxInt <span class="hljs-operator">=</span> <span class="hljs-keyword">type</span>(<span class="hljs-keyword">int</span>).<span class="hljs-built_in">max</span>;
 
-    <span class="hljs-keyword">address</span> <span class="hljs-keyword">public</span> addr = <span class="hljs-number">0xCA35b7d915458EF540aDe6068dFe2F44E8fa733c</span>;
+    <span class="hljs-keyword">address</span> <span class="hljs-keyword">public</span> addr <span class="hljs-operator">=</span> <span class="hljs-number">0xCA35b7d915458EF540aDe6068dFe2F44E8fa733c</span>;
 
     <span class="hljs-comment">/*
     In Solidity, the data type byte represent a sequence of bytes. 
@@ -55,8 +55,8 @@ const html = `<p>Here we introduce you to some primitive data types available in
      The term bytes in Solidity represents a dynamic array of bytes. 
      It’s a shorthand for byte[] .
     */</span>
-    <span class="hljs-keyword">bytes1</span> a = <span class="hljs-number">0xb5</span>; <span class="hljs-comment">//  [10110101]</span>
-    <span class="hljs-keyword">bytes1</span> b = <span class="hljs-number">0x56</span>; <span class="hljs-comment">//  [01010110]</span>
+    <span class="hljs-keyword">bytes1</span> a <span class="hljs-operator">=</span> <span class="hljs-number">0xb5</span>; <span class="hljs-comment">//  [10110101]</span>
+    <span class="hljs-keyword">bytes1</span> b <span class="hljs-operator">=</span> <span class="hljs-number">0x56</span>; <span class="hljs-comment">//  [01010110]</span>
 
     <span class="hljs-comment">// Default values</span>
     <span class="hljs-comment">// Unassigned variables have a default value</span>
