@@ -22,16 +22,16 @@ const html = `<p>There are 3 types of variables in Solidity</p>
 
 <span class="hljs-class"><span class="hljs-keyword">contract</span> <span class="hljs-title">Variables</span> </span>{
     <span class="hljs-comment">// State variables are stored on the blockchain.</span>
-    <span class="hljs-keyword">string</span> <span class="hljs-keyword">public</span> text = <span class="hljs-string">"Hello"</span>;
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> num = <span class="hljs-number">123</span>;
+    <span class="hljs-keyword">string</span> <span class="hljs-keyword">public</span> text <span class="hljs-operator">=</span> <span class="hljs-string">"Hello"</span>;
+    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> num <span class="hljs-operator">=</span> <span class="hljs-number">123</span>;
 
     <span class="hljs-function"><span class="hljs-keyword">function</span> <span class="hljs-title">doSomething</span>(<span class="hljs-params"></span>) <span class="hljs-title"><span class="hljs-keyword">public</span></span> </span>{
         <span class="hljs-comment">// Local variables are not saved to the blockchain.</span>
-        <span class="hljs-keyword">uint</span> i = <span class="hljs-number">456</span>;
+        <span class="hljs-keyword">uint</span> i <span class="hljs-operator">=</span> <span class="hljs-number">456</span>;
 
         <span class="hljs-comment">// Here are some global variables</span>
-        <span class="hljs-keyword">uint</span> timestamp = <span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span>; <span class="hljs-comment">// Current block timestamp</span>
-        <span class="hljs-keyword">address</span> sender = <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>; <span class="hljs-comment">// address of the caller</span>
+        <span class="hljs-keyword">uint</span> timestamp <span class="hljs-operator">=</span> <span class="hljs-built_in">block</span>.<span class="hljs-built_in">timestamp</span>; <span class="hljs-comment">// Current block timestamp</span>
+        <span class="hljs-keyword">address</span> sender <span class="hljs-operator">=</span> <span class="hljs-built_in">msg</span>.<span class="hljs-built_in">sender</span>; <span class="hljs-comment">// address of the caller</span>
     }
 }
 </code></pre>
